Countdown_Timer
A Countdown_Timer that counts down and when it hits zero it plays a sound also if the user is on the home screen it sends them a notifications. The goal  for this project is to learn about the notification system and to understand how I could use sounds in my apps.